+++ tests/test_ai_helper.py
@@ -1,4 +1,6 @@
 import unittest
+
+import utils.reference_validation as rv
 from utils.ai_helper import get_child_level_diagnoses_from_condition, get_inclusion_genomic_criteria, get_exclusion_genomic_criteria
 
 class TestAITasks(unittest.TestCase):
@@ -6,13 +8,7 @@ class TestAITasks(unittest.TestCase):
     def setUp(self):
         self.nct_condition = 'Colorectal Cancer'
         self.child_nodes_oncotree_list = ['Signet Ring Cell Adenocarcinoma of the Colon and Rectum', 'Colon Adenocarcinoma In Situ', 'Small Bowel Well-Differentiated Neuroendocrine Tumor', 'Gastrointestinal Neuroendocrine Tumors', 'Well-Differentiated Neuroendocrine Tumor of the Rectum', 'Small Bowel Cancer', 'Anal Squamous Cell Carcinoma', 'Anorectal Mucosal Melanoma', 'Low-grade Appendiceal Mucinous Neoplasm', 'Medullary Carcinoma of the Colon', 'Goblet Cell Adenocarcinoma of the Appendix', 'Mucinous Adenocarcinoma of the Appendix', 'Appendiceal Adenocarcinoma', 'Small Intestinal Carcinoma', 'Well-Differentiated Neuroendocrine Tumor of the Appendix', 'Signet Ring Cell Type of the Appendix', 'Colorectal Adenocarcinoma', 'High-Grade Neuroendocrine Carcinoma of the Colon and Rectum', 'Colonic Type Adenocarcinoma of the Appendix', 'Anal Gland Adenocarcinoma', 'Rectal Adenocarcinoma', 'Mucinous Adenocarcinoma of the Colon and Rectum', 'Duodenal Adenocarcinoma', 'Colon Adenocarcinoma', 'Tubular Adenoma of the Colon']
-        self.genes = self.get_gene_list()
-
-    def get_gene_list(self) -> list:
-        genes = []
-        with open('ref/genes.txt', 'r') as file:     
-            genes = [line.strip() for line in file.readlines()]
-        return genes
+        self.genes = sorted(rv.gene_symbols())
 
    
     def test_get_child_level_diagnoses_from_condition(self):